--- src/components/Sidebar.tsx
@@ -1,12 +1,0 @@
-import React, { FC } from "react";
-import { HiSearch } from "react-icons/hi";
-import { IoMdAddCircleOutline } from "react-icons/io";
-import { HiViewList } from "react-icons/hi";
-import { FiLogOut } from "react-icons/fi";
-import { NavLink } from "react-router-dom";
-
-const Sidebar = () => {
-  return <div>Sidebar </div>;
-};
-
-export default Sidebar;
