+++ tailwind.config.js
@@ -6,10 +6,7 @@ module.exports = {
     "./src/pages/Login.tsx",
     "./src/pages/Home.tsx",
     "./src/components/Button.tsx",
-    "./src/components/PasswordInput.component.tsx",
-    "./src/components/TextInput.component.tsx",
-    "./src/components/Sidebar.tsx",
-    "./src/pages/CreateBlog.tsx",
+    "./src/components/CustomInput.component.tsx",
     "./src/pages/MyBlogs.tsx",
     "./src/pages/ViewBlog.tsx",
   ],
